Move test to PostgresSqlModuleSpec
